+++ Tests/NewPlatform.Flexberry.ORM.ODataService.Tests/CRUD/Read/FilterByMasterDetailFieldTest.cs
@@ -66,7 +66,6 @@
                 DataObject[] newDataObjects = new DataObject[] { лес1, лес2, медведь1, медведь2, берлога1, берлога2, берлога3, берлога4, блоха1, блоха2, блоха3, блоха4 };
 
                 args.DataService.UpdateObjects(ref newDataObjects);
-                ExternalLangDef.LanguageDef.DataService = args.DataService;
 
                 string лес1Pk = ((KeyGuid)лес1.__PrimaryKey).Guid.ToString("D");
 
@@ -116,8 +115,6 @@
                 DataObject[] newDataObjects = new DataObject[] { лес1, лес2, медведь1, медведь2, берлога1, берлога2, берлога3, берлога4, блоха1, блоха2, блоха3, блоха4 };
 
                 args.DataService.UpdateObjects(ref newDataObjects);
-                ExternalLangDef.LanguageDef.DataService = args.DataService;
-
                 string лес1Pk = ((KeyGuid)лес1.__PrimaryKey).Guid.ToString("D");
                 string медведь1Pk = ((KeyGuid)медведь1.__PrimaryKey).Guid.ToString("D");
 
@@ -165,9 +162,7 @@
                 медведь3.Берлога.AddRange(берлога5, берлога6);
 
                 DataObject[] newDataObjects = new DataObject[] { лес1, лес2, лес3, медведь1, медведь2, берлога1, берлога2, берлога3, берлога4, берлога5, берлога6 };
-
                 args.DataService.UpdateObjects(ref newDataObjects);
-                ExternalLangDef.LanguageDef.DataService = args.DataService;
 
                 // Act.
                 string requestUrl = string.Format(
@@ -213,9 +208,7 @@
                 driver3.Car.AddRange(car1d3, car2d3);
 
                 DataObject[] newDataObjects = new DataObject[] { driver1, driver2, driver3, car1d1, car2d1, car1d2, car2d2, car1d3, car2d3 };
-
                 args.DataService.UpdateObjects(ref newDataObjects);
-                ExternalLangDef.LanguageDef.DataService = args.DataService;
 
                 // Act.
                 string requestUrl = string.Format(