+++ Tests/NewPlatform.Flexberry.ORM.ODataService.Tests/Startup.cs
@@ -5,10 +5,10 @@ namespace ODataServiceSample.AspNetCore
     using System.Linq;
     using ICSSoft.Services;
     using ICSSoft.STORMNET.Business;
+    using ICSSoft.STORMNET.Business.Audit;
+    using ICSSoft.STORMNET.Business.Interfaces;
     using ICSSoft.STORMNET.Security;
-    using ICSSoft.STORMNET.Windows.Forms;
     using IIS.Caseberry.Logging.Objects;
-    using Microsoft.AspNet.OData.Extensions;
     using Microsoft.AspNetCore.Builder;
     using Microsoft.AspNetCore.Hosting;
     using Microsoft.AspNetCore.Hosting.Server.Features;